Refactor bundle proof into focused module
